# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Letterboxd Wrapped API",
    "description": "Analyze Letterboxd exports and generate wrapped-style film statistics.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/feedback": {
      "post": {
        "summary": "Submit Feedback",
        "operationId": "submit_feedback_api_feedback_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_feedback_api_feedback_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report": {
      "post": {
        "summary": "Submit Report",
        "operationId": "submit_report_api_report_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_report_api_report_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_submit_feedback_api_feedback_post": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "general"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "include_names": {
            "type": "boolean",
            "title": "Include Names",
            "default": false
          },
          "attachment": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachment"
          }
        },
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "Body_submit_feedback_api_feedback_post"
      },
      "Body_submit_report_api_report_post": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "include_names": {
            "type": "boolean",
            "title": "Include Names",
            "default": false
          },
          "bundle": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Bundle"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "bundle"
        ],
        "title": "Body_submit_report_api_report_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}